Refactor page commands
