+++ eod/elements/pagecmds.go
@@ -2,10 +2,9 @@ package elements
 
 import (
 	"fmt"
-	"regexp"
 	"sort"
-	"strings"
 
+	"github.com/Nv7-Github/Nv7Haven/eod/eodsort"
 	"github.com/Nv7-Github/Nv7Haven/eod/types"
 	"github.com/Nv7-Github/Nv7Haven/eod/util"
 )
@@ -13,48 +12,50 @@ import (
 func (b *Elements) InvCmd(user string, m types.Msg, rsp types.Rsp, sorter string, filter string) {
 	rsp.Acknowledge()
 
-	b.lock.RLock()
-	dat, exists := b.dat[m.GuildID]
-	b.lock.RUnlock()
-	if !exists {
-		rsp.ErrorMessage("Guild not setup!")
-		return
-	}
-
-	inv, res := dat.GetInv(user, user == m.Author.ID)
+	db, res := b.GetDB(m.GuildID)
 	if !res.Exists {
 		rsp.ErrorMessage(res.Message)
 		return
 	}
-	items := make([]string, len(inv.Elements))
+
+	inv := db.GetInv(user)
+	items := make([]int, len(inv.Elements))
 	i := 0
-	dat.Lock.RLock()
+	db.RLock()
 	for k := range inv.Elements {
-		el, _ := dat.GetElement(k, true)
-		items[i] = el.Name
+		el, _ := db.GetElement(k, true)
+		items[i] = el.ID
 		i++
 	}
 
 	switch filter {
 	case "madeby":
 		count := 0
-		outs := make([]string, len(items))
+		outs := make([]int, len(items))
 		for _, val := range items {
 			creator := ""
-			elem, res := dat.GetElement(val, true)
+			elem, res := db.GetElement(val, true)
 			if res.Exists {
 				creator = elem.Creator
 			}
 			if creator == user {
-				outs[count] = val
+				outs[count] = elem.ID
 				count++
 			}
 		}
 		outs = outs[:count]
-		items = outs
 	}
-	util.SortElemList(items, sorter, dat)
-	dat.Lock.RUnlock()
+	eodsort.SortElemList(items, sorter, db)
+
+	text := make([]string, len(items))
+	for i, v := range items {
+		elem, res := db.GetElement(v, true)
+		if !res.Exists {
+			continue
+		}
+		text[i] = elem.Name
+	}
+	db.RUnlock()
 
 	name := m.Author.Username
 	if m.Author.ID != user {
@@ -66,29 +67,24 @@ func (b *Elements) InvCmd(user string, m types.Msg, rsp types.Rsp, sorter string
 	}
 	b.base.NewPageSwitcher(types.PageSwitcher{
 		Kind:       types.PageSwitchInv,
-		Title:      fmt.Sprintf("%s's Inventory (%d, %s%%)", name, len(items), util.FormatFloat(float32(len(items))/float32(len(dat.Elements))*100, 2)),
+		Title:      fmt.Sprintf("%s's Inventory (%d, %s%%)", name, len(items), util.FormatFloat(float32(len(items))/float32(len(db.Elements))*100, 2)),
 		PageGetter: b.base.InvPageGetter,
-		Items:      items,
+		Items:      text,
 	}, m, rsp)
 }
 
 func (b *Elements) LbCmd(m types.Msg, rsp types.Rsp, sorter string, user string) {
-	b.lock.RLock()
-	dat, exists := b.dat[m.GuildID]
-	b.lock.RUnlock()
-	if !exists {
-		return
-	}
-	_, res := dat.GetInv(user, user == m.Author.ID) // Check if user exists
+	db, res := b.GetDB(m.GuildID)
 	if !res.Exists {
 		rsp.ErrorMessage(res.Message)
 		return
 	}
+	db.GetInv(user) // Make user exist
 
 	// Sort invs
-	invs := make([]types.Inventory, len(dat.Inventories))
+	invs := make([]*types.Inventory, len(db.Invs()))
 	i := 0
-	for _, v := range dat.Inventories {
+	for _, v := range db.Invs() {
 		invs[i] = v
 		i++
 	}
@@ -127,98 +123,5 @@ func (b *Elements) LbCmd(m types.Msg, rsp types.Rsp, sorter string, user string)
 		Users:   users,
 		UserPos: userpos,
 		Cnts:    cnts,
-	}, m, rsp)
-}
-
-func (b *Elements) SearchCmd(search string, sort string, source string, opt string, regex bool, m types.Msg, rsp types.Rsp) {
-	b.lock.RLock()
-	dat, exists := b.dat[m.GuildID]
-	b.lock.RUnlock()
-	if !exists {
-		return
-	}
-	rsp.Acknowledge()
-	_, res := dat.GetInv(m.Author.ID, true)
-	if !res.Exists {
-		rsp.ErrorMessage(res.Message)
-		return
-	}
-
-	var list map[string]types.Empty
-	switch source {
-	case "elements":
-		list = make(map[string]types.Empty, len(dat.Elements))
-		for _, el := range dat.Elements {
-			list[el.Name] = types.Empty{}
-		}
-
-	case "inventory":
-		inv, res := dat.GetInv(opt, m.Author.ID == opt)
-		if !res.Exists {
-			rsp.ErrorMessage(res.Message)
-			return
-		}
-
-		list = make(map[string]types.Empty, len(inv.Elements))
-		dat.Lock.RLock()
-		for el := range inv.Elements {
-			elem, res := dat.GetElement(el, true)
-			if !res.Exists {
-				list[el] = types.Empty{}
-				continue
-			}
-			list[elem.Name] = types.Empty{}
-		}
-		dat.Lock.RUnlock()
-
-	case "category":
-		cat, res := dat.GetCategory(opt)
-		if !res.Exists {
-			rsp.ErrorMessage(res.Message)
-			return
-		}
-		list = cat.Elements
-	}
-
-	items := make(map[string]types.Empty)
-	if regex {
-		reg, err := regexp.Compile(search)
-		if rsp.Error(err) {
-			return
-		}
-		for el := range list {
-			m := reg.Find([]byte(el))
-			if m != nil {
-				items[el] = types.Empty{}
-			}
-		}
-	} else {
-		s := strings.ToLower(search)
-		for el := range list {
-			if strings.Contains(strings.ToLower(el), s) {
-				items[el] = types.Empty{}
-			}
-		}
-	}
-
-	txt := make([]string, len(items))
-	i := 0
-	for k := range items {
-		txt[i] = k
-		i++
-	}
-	util.SortElemList(txt, sort, dat)
-
-	if len(txt) == 0 {
-		rsp.Message("No results!")
-		return
-	}
-
-	b.base.NewPageSwitcher(types.PageSwitcher{
-		Kind:       types.PageSwitchInv,
-		Title:      fmt.Sprintf("Element Search (%d)", len(txt)),
-		PageGetter: b.base.InvPageGetter,
-		Items:      txt,
-		User:       m.Author.ID,
 	}, m, rsp)
 }
